Changed strto[l|ul|ull] functions to have correct signature
Updated man page as per Giacomo's comment

+++ sys/src/liblex/yyless.c
@@ -18,7 +18,7 @@ extern	int	yyprevious;
 void	yyunput(int c);
 
 void
-yyless(int x)
+yyless(uintptr_t x)
 {
 	char *lastch, *ptr;
